refactor(integration) resolve module path
