+++ integration/graphql/e2e/graphql-async.spec.ts
@@ -1,7 +1,7 @@
 import { INestApplication } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import * as request from 'supertest';
-import { AsyncApplicationModule } from './../src/async-options.module';
+import { AsyncApplicationModule } from '../src/async-options.module';
 
 describe('GraphQL (async configuration)', () => {
   let app: INestApplication;
