+++ integration/hello-world/e2e/exceptions.spec.ts
@@ -1,7 +1,6 @@
 import * as request from 'supertest';
 import { Test } from '@nestjs/testing';
 import { INestApplication, HttpStatus } from '@nestjs/common';
-import { ApplicationModule } from './../src/app.module';
 import { ErrorsController } from '../src/errors/errors.controller';
 
 describe('Error messages', () => {
