+++ integration/hello-world/e2e/express-instance.spec.ts
@@ -2,7 +2,7 @@ import * as express from 'express';
 import * as request from 'supertest';
 import { Test } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
-import { ApplicationModule } from './../src/app.module';
+import { ApplicationModule } from '../src/app.module';
 
 describe('Hello world (express instance)', () => {
   let server;
