+++ integration/hello-world/e2e/guards.spec.ts
@@ -1,7 +1,7 @@
 import * as request from 'supertest';
 import { Test } from '@nestjs/testing';
 import { INestApplication, Injectable, UnauthorizedException } from '@nestjs/common';
-import { ApplicationModule } from './../src/app.module';
+import { ApplicationModule } from '../src/app.module';
 import { APP_GUARD } from '@nestjs/core';
 
 @Injectable()
