+++ integration/hello-world/e2e/interceptors.spec.ts
@@ -1,7 +1,7 @@
 import * as request from 'supertest';
 import { Test } from '@nestjs/testing';
 import { INestApplication, Injectable } from '@nestjs/common';
-import { ApplicationModule } from './../src/app.module';
+import { ApplicationModule } from '../src/app.module';
 import { APP_INTERCEPTOR } from '@nestjs/core';
 import { of } from 'rxjs';
 import { map } from 'rxjs/operators';
