+++ integration/typeorm/e2e/typeorm-async-existing.spec.ts
@@ -1,7 +1,7 @@
 import { INestApplication } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import * as request from 'supertest';
-import { AsyncOptionsExistingModule } from './../src/async-existing-options.module';
+import { AsyncOptionsExistingModule } from '../src/async-existing-options.module';
 
 describe('TypeOrm (async configuration)', () => {
   let server;
