+++ integration/typeorm/e2e/typeorm-async-options.spec.ts
@@ -1,7 +1,7 @@
 import { INestApplication } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import * as request from 'supertest';
-import { AsyncOptionsFactoryModule } from './../src/async-options.module';
+import { AsyncOptionsFactoryModule } from '../src/async-options.module';
 
 describe('TypeOrm (async configuration)', () => {
   let server;
