+++ integration/typeorm/e2e/typeorm-async.spec.ts
@@ -1,7 +1,7 @@
 import { INestApplication } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import * as request from 'supertest';
-import { AsyncApplicationModule } from './../src/app-async.module';
+import { AsyncApplicationModule } from '../src/app-async.module';
 
 describe('TypeOrm (async configuration)', () => {
   let server;
